move IsPkceClient extension to the host project
