+++ src/IdentityServer4/Extensions/IClientStoreExtensions.cs
@@ -25,22 +25,5 @@ namespace IdentityServer4.Stores
 
             return null;
         }
-
-        /// <summary>
-        /// Determines whether the client is configured to use PKCE.
-        /// </summary>
-        /// <param name="store">The store.</param>
-        /// <param name="client_id">The client identifier.</param>
-        /// <returns></returns>
-        public static async Task<bool> IsPkceClientAsync(this IClientStore store, string client_id)
-        {
-            if (!string.IsNullOrWhiteSpace(client_id))
-            {
-                var client = await store.FindEnabledClientByIdAsync(client_id);
-                return client?.RequirePkce == true;
-            }
-
-            return false;
-        }
     }
 }
